Platform polish: dismiss/retry semantics fixes, worker core split, typecheck + tests
Bug fixes (user-reported via console testing):
- Retry/re-open of a summarized link now REUSES its NNN (was: always
  allocated a fresh ID → duplicate summary + counter double-spend)
- Dismissing a summarized link now RETRACTS its workdesk summary (was:
  orphaned summaries would leak to consumers); published rows never
  touched; NNN stays spent; console confirms before retracting
- DO final status updates guarded on status='queued' so a dismissal
  racing an in-flight run wins

Refactors (as promised for the polish PR):
- worker/src split: index.ts (DO + routes) / core.ts (summarizeUrl
  pipeline core shared with /eval) — no more mixed concerns
- tsc --noEmit type-checking wired (platform/package.json + tsconfig,
  strict; passes) and vitest unit tests for the pure logic (9 tests:
  sanitize rules, URL validation, summary/stub classification, token
  usage parsing)
- platform/README.md rewritten to match reality (layout, commands,
  auth model, the semantics worth remembering)

Production: orphaned summaries 001/002 (from dismissed links) removed;
e2e verified post-refactor — submit → summarize → dismiss → retraction
confirmed, counter restored.
