+++ platform/functions/api/links/[id].ts
@@ -26,14 +26,26 @@ export const onRequestPatch: PagesFunction<Env> = async ({ request, env, params,
 
   const consumedAt = status === "consumed" ? new Date().toISOString() : null;
   const res = await env.DB.prepare(
-    "UPDATE links SET status = ?, consumed_at = ?, error = CASE WHEN ? = 'new' THEN NULL ELSE error END WHERE id = ? RETURNING id, url, status, consumed_at",
+    "UPDATE links SET status = ?, consumed_at = ?, error = CASE WHEN ? = 'new' THEN NULL ELSE error END WHERE id = ? RETURNING id, url, status, consumed_at, summary_id",
   )
     .bind(status, consumedAt, status, id)
-    .first();
+    .first<{ id: number; url: string; status: string; consumed_at: string | null; summary_id: string | null }>();
   if (!res) return error("link not found", 404);
+
+  let retracted: string | null = null;
+  if (status === "dismissed" && res.summary_id) {
+    // Dismiss means "not in the journal": retract the link's workdesk summary.
+    // Published rows are never touched; the NNN stays spent (gaps are honest).
+    const del = await env.DB.prepare(
+      "DELETE FROM summaries WHERE id = ? AND journal_date IS NULL AND status = 'workdesk' RETURNING id",
+    )
+      .bind(res.summary_id)
+      .first();
+    if (del) retracted = res.summary_id;
+  }
   if (status === "new") {
     const kick = enqueueSummarization(env);
     if (kick) waitUntil(kick);
   }
-  return json(res);
+  return json({ ...res, retracted_summary: retracted });
 };
